requirements.txt
requests>=2.31.0
openpyxl>=3.1.2
streamlit>=1.30.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
PyGithub>=2.1.0
